Add description to output, mark more inputs as optional
After publishing to Terraform's registry, the auto-generated docs
had two optional variables listed as required. Hopefully, this explicit
change to an empty string value will change that.

I'm not sure if there is a difference betweeen passing `null` or `""`
to the `name_prefix` so I jump through a few function hoops to directly
pass what the module caller sets.